+++ google/src/test/scala/org/broadinstitute/dsde/workbench/google/GoogleUtilitiesSpec.scala
@@ -7,7 +7,7 @@ import akka.testkit.TestKit
 import com.google.api.client.googleapis.json.GoogleJsonError.ErrorInfo
 import com.google.api.client.googleapis.json.{GoogleJsonError, GoogleJsonResponseException}
 import com.google.api.client.http._
-import org.broadinstitute.dsde.workbench.metrics.StatsDTestUtils
+import org.broadinstitute.dsde.workbench.metrics.{Histogram, StatsDTestUtils}
 import org.broadinstitute.dsde.workbench.util.MockitoTestUtils
 import org.scalatest.{BeforeAndAfterAll, FlatSpecLike, Matchers}
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
@@ -19,9 +19,9 @@ import scala.concurrent.duration._
 
 class GoogleUtilitiesSpec extends TestKit(ActorSystem("MySpec")) with GoogleUtilities with FlatSpecLike with BeforeAndAfterAll with Matchers with ScalaFutures with Eventually with MockitoTestUtils with StatsDTestUtils {
   implicit val executionContext = ExecutionContext.global
-  implicit def histo = ExpandedMetricBuilder.empty.asHistogram("histo")
+  implicit def histo: Histogram = ExpandedMetricBuilder.empty.asHistogram("histo")
 
-  override def afterAll {
+  override def afterAll: Unit = {
     TestKit.shutdownActorSystem(system)
   }
 
@@ -97,7 +97,7 @@ class GoogleUtilitiesSpec extends TestKit(ActorSystem("MySpec")) with GoogleUtil
   "retryWhen500orGoogleError" should "retry once per backoff interval and then fail" in {
     withStatsD {
       val counter = new Counter()
-      whenReady(retryWhen500orGoogleError(counter.alwaysBoom).failed) { f =>
+      whenReady(retryWhen500orGoogleError(() => counter.alwaysBoom()).failed) { f =>
         f shouldBe a[IOException]
         counter.counter shouldBe 4 //extra one for the first attempt
       }
@@ -110,7 +110,7 @@ class GoogleUtilitiesSpec extends TestKit(ActorSystem("MySpec")) with GoogleUtil
   it should "not retry after a success" in {
     withStatsD {
       val counter = new Counter()
-      whenReady(retryWhen500orGoogleError(counter.boomOnce)) { s =>
+      whenReady(retryWhen500orGoogleError(() => counter.boomOnce())) { s =>
         s shouldBe 42
         counter.counter shouldBe 2
       }
@@ -128,7 +128,7 @@ class GoogleUtilitiesSpec extends TestKit(ActorSystem("MySpec")) with GoogleUtil
         case _: IOException => 42
       }
 
-      whenReady(retryWithRecoverWhen500orGoogleError(counter.alwaysBoom)(recoverIO)) { s =>
+      whenReady(retryWithRecoverWhen500orGoogleError(() => counter.alwaysBoom())(recoverIO)) { s =>
         s shouldBe 42
         counter.counter shouldBe 1
       }
@@ -146,7 +146,7 @@ class GoogleUtilitiesSpec extends TestKit(ActorSystem("MySpec")) with GoogleUtil
         case h: HttpResponseException if h.getStatusCode == 404 => 42
       }
 
-      whenReady(retryWithRecoverWhen500orGoogleError(counter.httpBoom)(recoverHttp).failed) { f =>
+      whenReady(retryWithRecoverWhen500orGoogleError(() => counter.httpBoom())(recoverHttp).failed) { f =>
         f shouldBe a[HttpResponseException]
         counter.counter shouldBe 4 //extra one for the first attempt
       }